refactored: endpoint models
