--- srcs/models/__init__.py
@@ -1,7 +1,0 @@
-from models.search import *
-from models.get_embedding import *
-from models.get_metadata import *
-from models.get_metadata_with_embedding import *
-from models.memo import *
-from models.memo import *
-from models.kakao_parser import *
