--- srcs/models/memo/__init__.py
@@ -1,4 +1,0 @@
-from models.memo.memo import *
-from models.memo.structures import *
-from models.memo.tag import *
-from models.memo.tags import *
