+++ srcs/routers/_models/embedding/get_embedding.py
@@ -1,7 +1,7 @@
 from pydantic import BaseModel
 
 
-class Arg_get_embedding(BaseModel):
+class Body_get_embedding(BaseModel):
     content: str
 
 class Res_get_embedding(BaseModel):
