+++ srcs/routers/_models/embedding/get_metadata.py
