+++ srcs/routers/_models/embedding/get_metadata_with_embedding.py
