+++ srcs/routers/_models/memo/_models/tag.py
@@ -1,0 +1,9 @@
+from typing import Optional
+from pydantic import BaseModel
+
+    
+class Memo_tag(BaseModel):
+    id: str
+    name: str
+    is_new: bool
+    embedding: Optional[list[float]]=None
