+++ srcs/routers/_models/memo/_models/tag_name_and_id.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+    
+class Memo_tag_name_and_id(BaseModel):
+    id: str
+    name: str
+    is_new: bool
