+++ srcs/routers/_models/parser/_models/__init__.py
@@ -1,0 +1,1 @@
+from .kakao_parser_type import Kakao_parser_type
