+++ srcs/routers/_models/parser/_models/kakao_parser_type.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+
+class Kakao_parser_type(Enum):
+    CSV="csv"
+    TXT="txt"
