+++ srcs/routers/_models/parser/kakao_parser.py
@@ -1,11 +1,8 @@
 from pydantic import BaseModel, Field
-from enum import Enum
+from routers._models.parser._models import Kakao_parser_type
 
-class kakao_parser_type(Enum):
-    CSV="csv"
-    TXT="txt"
 
 class Body_post_kakao_parser(BaseModel):
     user_id: str
-    type: kakao_parser_type
+    type: Kakao_parser_type
     content: str=Field(description="url of content")
