+++ srcs/routers/_models/search/__init__.py
@@ -1,0 +1,2 @@
+from ._models import *
+from .search import *
