+++ srcs/routers/_models/search/_models/__init__.py
@@ -1,0 +1,1 @@
+from .search_type import Search_query_type
