+++ srcs/routers/_models/search/_models/search_type.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+
+class Search_query_type(Enum):
+    regex = "regex"
+    similarity = "similarity"
